refactor: extract all common code from syn to a single file
