+++ znap-syn/src/common.rs
@@ -1,4 +1,24 @@
-use syn::{FnArg, GenericArgument, Ident, ItemFn, PathArguments, ReturnType, Type};
+use deluxe::parse2;
+use syn::{Ident, ItemStruct, FnArg, GenericArgument, ItemFn, PathArguments, ReturnType, Type};
+use heck::{ToSnekCase, ToUpperCamelCase};
+use proc_macro2::Span;
+
+pub fn extract_query_type(action_struct: &ItemStruct) -> Option<Ident> {
+    action_struct.attrs.iter().find_map(|attr| {
+        if let Ok(meta) = attr.meta.require_list() {
+            if let Some(first_segment) = meta.path.segments.first() {
+                if first_segment.ident.to_string() == "query" {
+                    if let Ok(query) = parse2::<Ident>(meta.tokens.clone()) {
+                        return Some(query);
+                    }
+                }
+            }
+        }
+
+        return None;
+    })
+}
+
 
 pub fn extract_action_ident(f: &ItemFn) -> Option<&Ident> {
     if let FnArg::Typed(pt) = f.sig.inputs.first()? {
@@ -62,4 +82,69 @@ pub fn extract_fn_result_type(f: &ItemFn) -> Option<&Ident> {
         }
     }
     None
+}
+
+pub fn create_query(action: &String) -> Ident {
+    Ident::new(
+        &format!(
+            "{}Query",
+            action_name_without_suffix(&action.to_snek_case()).to_upper_camel_case()
+        ),
+        Span::call_site(),
+    )
+}
+
+pub fn create_get_handler(action: &String) -> Ident {
+    Ident::new(
+        &format!("handle_get_{}", action_name_without_suffix(&action.to_snek_case())),
+        Span::call_site(),
+    )
+}
+
+pub fn create_post_handler(action: &String) -> Ident {
+    Ident::new(
+        &format!("handle_post_{}", action_name_without_suffix(&action.to_snek_case())),
+        Span::call_site(),
+    )
+}
+
+pub fn create_route_path(action: &String) -> String {
+    format!(
+        "/api/{}",
+        action_name_without_suffix(&action.to_snek_case()).to_snek_case()
+    )
+}
+
+pub fn create_post_context(action: &String) -> Ident {
+    Ident::new(
+        &format!(
+            "{}PostContext",
+            action_name_without_suffix(&action.to_snek_case()).to_upper_camel_case()
+        ),
+        Span::call_site(),
+    )
+}
+
+pub fn create_transaction(action: &String) -> Ident {
+    Ident::new(
+        &format!("{}_create_transaction", action_name_without_suffix(&action.to_snek_case())),
+        Span::call_site(),
+    )
+}
+
+pub fn create_metadata(action: &String) -> Ident {
+    Ident::new(
+        &format!("{}_create_metadata", action_name_without_suffix(&action.to_snek_case())),
+        Span::call_site(),
+    )
+}
+
+pub fn create_get_context(action: &String) -> Ident {
+    Ident::new(
+        &format!(
+            "{}GetContext",
+            action_name_without_suffix(&action.to_snek_case()).to_upper_camel_case()
+        ),
+        Span::call_site(),
+    )
 }
