+++ znap-syn/src/parser/action/attributes.rs
@@ -1,7 +1,6 @@
 use deluxe::extract_attributes;
 use quote::ToTokens;
 use syn::{parse::Result as ParseResult, parse2, ItemStruct};
-
 use crate::ActionAttributesStruct;
 
 pub fn parse(action_struct: &ItemStruct) -> ParseResult<ActionAttributesStruct> {
