+++ znap-syn/src/parser/action/mod.rs
@@ -1,14 +1,16 @@
 use crate::ActionStruct;
 use syn::{parse::Result as ParseResult, ItemStruct};
-
 mod attributes;
+mod query;
 
 pub fn parse(action_struct: &ItemStruct) -> ParseResult<ActionStruct> {
     let action_attributes = attributes::parse(&action_struct)?;
+    let query = query::parse(&action_struct);
 
     Ok(ActionStruct {
         name: action_struct.ident.clone(),
         raw_struct: action_struct.clone(),
         attributes: action_attributes,
+        query
     })
 }
