+++ znap-syn/src/parser/collection/actions.rs
@@ -1,10 +1,9 @@
-use super::common::extract_fn_result_type;
-use crate::parser::collection::common::extract_action_ident;
 use syn::{
     parse::{Error as ParseError, Result as ParseResult},
     spanned::Spanned,
     Ident, Item, ItemFn, ItemMod,
 };
+use crate::common::{extract_action_ident, extract_fn_result_type};
 
 pub fn parse(collection_mod: &ItemMod) -> ParseResult<Vec<Ident>> {
     let mod_content = &collection_mod
